Use SIESTA_CSC format in case of ELSI sparse interface.
It is a requirement of PEXSI (and PEXSI_CSC) that input matrices are
distributed across MPI tasks in charge of the first pole, whereas
SIESTA_CSC is more flexible.

Technically SIESTA_CSC is a 1D block-cyclic distribution. Here the
block size is specifically chosen to make it a block distribution.