Write tests for dp_bands and minor changes
Add test_dp_bands.py, as well as the corresponding reference databases,
for automated testing. Write equals-method for nxy-files in common.py
as well as the nxy-class. Move functionality of dp_bands to
../dptools/scrips/ to enable function call via the python package
dptools and add a command line wrapper in ../dptools/bin/. Add invalid-
input-file test for straingen and fix a bug where temporary test-folders
are not deleted in test_xyz2gen.py.