Replace use of allocation status by optionals
The use of the allocation status to check for presence of variables
has been kept as a workaround to maintain compatibility with some
compilers until this point. Since the corresponding compiler versions
no longer have to be supported, a change to the consistent use of
optional arguments is possible. This is a first step to accomplish it.