Symmetrises the charges in contact calculations
As the two principal layers are identical, their charges should
be symmetric. In the absence of code changes to just calculate
one PL and duplicate it, this instead symmetrises the charge
between equivalent atoms in the two layers.

Note, this will need revisiting for hybrids, but covers atomic,
block and multipolar populations.