Fix a not_allocated_bug in reksinterface
The reks%rhoSqrL can be used only when force calculation is true.
Otherwise, reks%rhoSpL is used in the code. In current reksEnProperties
routine, reks%rhoSqrL always passes to getUnrelaxedDensMatAndTdp routine
so that the non_allocated error occurs only when force calculation
is false. To solve this, I separate the variable into rhoSqrL and rhoSpL.