Parallelize atomic neighbour computation
The distance information of all atomic neighbours are needed in every
MPI task. The arrays were previously computed serially and stored on
each task. This caused out-of-memory errors for setups with many atoms
(at least several thousands). With this commit, the computation is
parallelized among all MPI tasks on a node and the arrays are stored
only once per node using MPI shared memory.