Symmetric neighbours with Born solvents
Use the symmetric neighbour list for the generalised Born solvent
model (as that internally assumes symmetric neighbours).

Note, seems invariant to permutation of atom order, but untested
whether the results are numerically correct.

Derives from the periodic hybrid functional branch, so will need to be
rebased and merged after it becomes incorporated.