Symmetric neighbours with solvents
Use the symmetric neighbour list for the solvent models
(as that part of the code internally assumes symmetric
neighbours, not the default single triangle, leading to
the energy being atomic order dependent).

Note, results now seem invariant to permutation of atom order,
but are untested whether the results are physically correct.