Checks consistency of cmake and submodule repo
Add flag for consistency check, as the github action was failing on
the script (which had no problems locally).

Two of the submodules were pointing to a fork or a non-existing fork
(github.com/dftbplus/chimes_calculator gives a 404). Also change to
the headless github repo URLs for various projects without the .git