Fixes array sizing/unallocated access
Changes:

The sparse overlap matrix sizing is only increased, so the call to get
sizing for the sparse density matrix is incorrect if the sizing
shrinks (overlap & H are only incremented to reduce allocate
operations).

Remove unused getSparseSize function.

The old delta density matrix is only allocated on certain code paths,
so should only be zeroed if allocated in those code paths.

The copy to allocate for the dRhoPrevCplxHS would happen twice if not
this%tScreeningInited.