Use toml for the ford settings
With the introduction of TOML support in Ford version 7.0, it is now
recommended to specify the Project Options in the [extra.ford] section
of a fpm.toml file.

This change migrates the settings from the markdown project file to an equivalent TOML configuration.
The Ford invocation remains unchanged and continues to simply be:
$ ford dftbplus-project-file.md

To avoid having the main project file be completely devoid of content, I copied over some material from the project README.