Serial XLBOMD solution least squares
Switch from inversion of overlap to solving A X = B instead, using
posv call, which assumes A is positive definite. But this is for the
purpose of Choleskii factoring, hence as the generalised eigenproblem
use this already to get to this point (so would fail first), it should
be stable in this case.