Minor regression test issues
As the SK set used in the rest of the tests is the shift not base
version and also there is a keyword change for hybrid functionals
again requiring explicit scc.

Note, three other tests are still not passing, partly from probable
change in tag name, but also quantities about currents.